Add neorg plugin to organise notes
This is a trial, I'm not sure what will happen with the plugin (it's
relatively young plugin).
But I really like the way it works compared to vimwiki. And there I have
to note that I'm not familiar enough with vimwiki, and that I did not
give it a try with its wiki format, but the rather undersupported
markdown format.